s3ext: remove dead codes before refactoring gpreader
1. add gpcommon.h to keep 'GpId' data structures;
2. remove s3reader.h because s3key_reader is ready;
3. move two pthread related functions to gpreader.h
   because they are invoked by gpreader only.
4. move XMLParserCallback and XMLInfo to s3uploader.cpp
   because they are deprecated already for readers.
