+++ gpAux/extensions/gps3ext/include/gpcheckcloud.h
@@ -10,7 +10,6 @@
 #include "s3conf.h"
 #include "s3key_reader.h"
 #include "s3log.h"
-#include "s3thread.h"
 
 #define BUF_SIZE 64 * 1024
