+++ gpAux/extensions/gps3ext/include/gpcommon.h
