+++ gpAux/extensions/gps3ext/include/s3common.h
@@ -8,6 +8,7 @@
 #include <libxml/parser.h>
 #include <libxml/tree.h>
 
+#include "gpcommon.h"
 #include "http_parser.h"
 #include "s3http_headers.h"
 #include "s3log.h"
@@ -23,12 +24,6 @@ enum Method { GET, PUT, POST, DELETE, HEAD };
 
 void SignRequestV4(const string& method, HTTPHeaders* h, const string& orig_region,
                    const string& path, const string& query, const S3Credential& cred);
-
-struct XMLInfo {
-    xmlParserCtxtPtr ctxt;
-};
-
-uint64_t XMLParserCallback(void* contents, uint64_t size, uint64_t nmemb, void* userp);
 
 char* get_opt_s3(const char* url, const char* key);
