--- gpAux/extensions/gps3ext/include/s3thread.h
@@ -1,8 +1,0 @@
-#ifndef __S3_THREAD_H__
-#define __S3_THREAD_H__
-
-int thread_setup(void);
-
-int thread_cleanup(void);
-
-#endif
