+++ gpAux/extensions/gps3ext/src/gps3ext.cpp
@@ -18,11 +18,9 @@
 #include "utils/memutils.h"
 
 #include "gpreader.h"
-#include "gps3ext.h"
 #include "s3common.h"
 #include "s3conf.h"
 #include "s3log.h"
-#include "s3thread.h"
 #include "s3utils.h"
 
 /* Do the module magic dance */
