+++ gpAux/extensions/gps3ext/src/s3conf.cpp
@@ -12,7 +12,6 @@
 #include <sstream>
 #include <string>
 
-#include "gps3ext.h"
 #include "s3conf.h"
 #include "s3log.h"
 #include "s3utils.h"
