+++ gpAux/extensions/gps3ext/src/s3extbase.cpp
@@ -1,7 +1,6 @@
 #include <sstream>
 #include <string>
 
-#include "gps3ext.h"
 #include "s3conf.h"
 #include "s3extbase.h"
 #include "s3log.h"
