+++ gpAux/extensions/gps3ext/test/gpreader_test.cpp
@@ -111,3 +111,14 @@ TEST(ExtBase, ValidateURL_apnortheast21) {
 
     delete myData;
 }
+
+TEST(Common, ThreadFunctions) {
+    // just to test if these two are functional
+    thread_setup();
+    EXPECT_NE((void *)NULL, mutex_buf);
+
+    thread_cleanup();
+    EXPECT_EQ((void *)NULL, mutex_buf);
+
+    EXPECT_NE(0, id_function());
+}
